backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pypdf==4.1.0
PyMuPDF==1.23.26
Pillow==10.2.0
aiofiles==23.2.1
